aws-playground/requirements.txt
boto3
python-dotenv
pinecone
sentence-transformers
langchain
langchain-community
